# polyglot/packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Полиглот — vocabulary trainer API with session-based spaced repetition
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tags:
    get:
      operationId: listTags
      tags: [tags]
      summary: List all custom tags optionally filtered by kind
      parameters:
        - in: query
          name: kind
          schema:
            type: string
          description: "mnemonic_group | semantic_group | word_type"
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagListResponse"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create a new custom tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /tags/{id}:
    put:
      operationId: updateTag
      tags: [tags]
      summary: Update a tag label or sort order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpdate"
      responses:
        "200":
          description: Updated tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "404":
          description: Tag not found
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete a custom tag
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Tag not found
components:
  schemas:
...
    Tag:
      type: object
      required: [id, kind, value, label, sortOrder]
      properties:
        id:
          type: integer
        kind:
          type: string
        value:
          type: string
        label:
          type: string
        sortOrder:
          type: integer
    TagListResponse:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
    TagInput:
      type: object
      required: [kind, value, label]
      properties:
        kind:
          type: string
        value:
          type: string
        label:
          type: string
        sortOrder:
          type: integer
    TagUpdate:
      type: object
      properties:
        label:
          type: string
        sortOrder:
          type: integer
...